+++ fin/ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for boring
 
+## 0.0.2
+
+- In `Fin` add: `append` and `split`
+- Add `(Enum a, Enum b) => Enum (Either a b)` instance
+
 ## 0.0.1
 
 - GHC-8.4.1 / base-4.11 support